Upgrade to PS 0.12
This is only partially complete.  Some dependencies have not yet been fully published.

fromCharCode, on which the parser relies heavily, has changed insofar as it now returns Maybe Char rather than Char.  I think as a consequence of this, the QuickCheck test for round-trip generation and parsing of a Midi Recording would occasionally fail. To mitigate this, I altered the generation of strings destined for Midi Events that are Strings to be restricted to printable ASCII characters.

The examples have not been updated.