Upgrade dependency on purescript-string-parsers
This is now the actual 0.12 release of string-parsers.  It has a breaking change whereby all the various many1 functions return NonEmptyList rather than List.  At the moment, this just affects SysEx events - we could possibly add others where the event is a counted array of bytes and the count should not be 0.