Reduce code bloat in Midi.Instruments
Huge code bloat seems to happen whenever there are a large number of
simple constructors in an ADT representing a simple enumerated type.
The bloat seems to be caused by:

* deriving Ord instances (enormous impact)
* deriving Eq instances (IncompleteExhaustivityCheck warning kicks in
after 100 constructors)
* Considerable bloat (about 300K) just by virtue of using the Unicode
functions toLower, toUpper and is Digit.

The first two issues also have a huge effect on compile time.

Fixed by hand-writing Eq and Ord instances and by using (trivial and
inefficient) versions of the 3 character functions because luckily, we
know all Gleitzman file names are in ASCII)